Break out components in LeaderBoard
