+++ client/src/containers/index.js
@@ -1,4 +1,4 @@
 export { default as Home } from './Home';
 export { default as About } from './About';
 export { default as Header } from './Header';
-export { default as LeaderBoard } from './LeaderBoard';
+//export { default as LeaderBoard } from './LeaderBoard';
